Checking input files for duplicate chemgraphs: Update
MRH has replaced the previous method of checking for dupliate chemgraphs in
any input file with a different method.

Old method: Check in the Species.make() function against all species
already present in the SpeciesDictionary.

Problem: If running a RMG job with Restart turned on, there is a good chance
the same species will be in the Restart coreSpecies.txt file AND the
user-specified input file.  This would cause RMG to fail.

New method: Within each class, have an independent LinkedHashMap that
stores all chemgraphs read in from the specific input file (i.e. do not
include the list of species from Restart).  Pass this LinkedHashMap and
the current read-in chemgraph to a function that checks the current
ChemGraph against all previously read-in ChemGraphs.  If not present, add to
list and continue.  If present, terminate immediately with error message.