Remove the (double)getRateConstant(_) methods that cause confusion.
These methods used to do the wrong thing (i.e. return a stale cached number).
They are not needed anyway, as all they do is call calculateTotalRate, which
is a (slightly) better name because it indicates that the answer is not a
constant. (It's still a confusing name because it's a rate coefficient, not a rate).
With this commit, the getRateConstant(Temperature) and
getRateConstant(Temperature,Pressure) methods have gone, and calls to them are
replaced with calculateTotalRate and calculateTotalPDepRate respectively.