PLOG pressures could be different for each instance of PDepArrheniusKinetics
Instead of being a static list of Pressures (and static int numPressures) these
are not static, i.e. they are stored on each instance of PDepArrheniusKinetics.
This allows each PDepArrheniusKinetics object to have a different list of pressures.
This is important for rates that are read in from reaction libraries or seed
mechanisms.