Mark seed mechanism reactions as being FromPrimaryKineticLibrary
It seems that this boolean attribute is used to decide whether the kinetics
should be multiplied by the reaction path degeneracy or left alone. The
idea being that things from a primary kinetic library were intended to be
used "as is" and without further multiplication. This is the case for
seed mechanisms, and so I am setting it for all reactions read in from
a SeedMechanism reader.  (Hopefully the Restart features which use the
SeedMechanisme.read function have written the multiplied kinetics?)


NB. This commit 'fixes' it in an ugly way - making all calls to the read[Pdep]Reactions()
method pass a 'true' boolean for the 'pkl' parameter. However, the nicer
fix - remove the parameter - which is is done on the seedmech branch, requires
some refactoring of the underlying code.