Throw POutOfRangeException when evaluating a PDepArrheniusKinetics outside valid P range.
I think what CHEMKIN does is just use the edge value, i.e. if you're below
the lowest use the lowest, and if you're above the highest use the highest.
This could be a good way to go because many of Glarborg's rates only go down
to 1 atm.  Before this commit however, it would just return a rate of 0 without
any warning or explanation!