Fix bug and improve reporting of RATE-formatted P-dep rates.
One mistake was introduced in f9518240b1f66d90690eddf4a3e142b33a5b1f36.

There was an additional mistake (I think) in that it was adding (+m), but
these rate coefficients are, as far as Chemkin is concerned, just a plain
old function of temperature. This mistake was older.

I have also added a comment about the conditions at which the rate was evaluated.