Methods to add PDepArrheniusKinetics for duplicate pathways.
You can add a PDepArrheniusKinetics to a PDepReaction using add PDepArrheniusKinetics(k)
which calls the method of the same name on its PDepRateConstant object. This creates
a new entry in the array (or makes an array if none exists).
Will be useful for when seed mechanisms declare things as the sum of duplicate PLOG reactions.