ThermoDataEstimator: Make user specify name for each chemgraph
Instead of listing only chemgraphs in the input file for the
ThermoDataEstimator class, force the user to also list a name with each
chemgraph.  This is consistent with all other RMG (and related class)
input files.

Example: A user has a list of InChIs.  Passing this list to InChI2AdjList
produces a list of names+chemgraphs.  If the user then wanted to know
the thermo or transport of each of these species, they can use the generated
file (with minor updates, e.g. adding PrimaryThermoLibrary field).  Before,
the user would need to remove every species name.