Instead of increasing redundancy, add A factors when you find multiple kinetics the same.
Previously, whenever kinetics were added to a reaction but the same kinetics were already
there (eg. you had already found an equivalent pathway with differently labeled atoms)
the redundancy *of the reaction structure* was increased. THis redundancy, however, applies
to all the pathways of that reaction. If there are two pathways, each is found several times,
the redendancy is incremented a lot, then multiplied by each pathway.

For the 1,3_Insertion_CO2 reaction: CO2 + H2 --> CH2O2
this would lead to final rates that were 16 times higher than the group values (instead of ~4?)

Now, if we find kinetics with the same N and E and Source (matched nodes) then we
add the A values.

I believe this closes #156.