Bug fix in UncertainDouble.getAddingUncertainty()
This function converts whatever form the uncertainty is in, into an additive one.
The trouble is, if your mulciplicative uncertainty is greater than (or equal to) 2,
then the additive uncertainty will be greater than (or equal to) the value, and
your lower bound will be negative. This is usually best avoided.
But with this fix, at least it does what it claims!