Improved means of adding with UncertainDouble.plus
Instead of forcing all the uncertainties to be additive (which has problems with
negative lower bounds), if both uncertainties were already multiplicative,
it leaves them as so, and uses the highest (or only if they're equal) uncertainty.
This makes the new add-redundant-A-factors-together code work much better, as
you can avoid having negative lower bounds on your A factors.