Force "dup" tag to be used in kinetic_libraries reactions.txt files
When a user supplies a kinetic library, reaction library, or seed mechanism:
RMG reads in the reactions.txt file.  If it finds more than one instance of
a reaction string (e.g. A+B=C), it used to assume that the user
wanted this duplicate instance and would add the kinetics of the
duplicate instance to the array of kinetics of the original instance
of the reaction.

Now, I am enforcing the user to specify "DUP" or "DUPLICATE" after each
duplicate reaction.  This should lead to fewer possible errors when
supplying a kinetics_library folder to RMG.

After reading in all reactions in the reactions.txt file, RMG now iterates over
all reactions.  If the reaction is marked as a duplicate, but has only one
set of ArrheniusKinetics, RMG will print an error message (with the offending
reaction) and terminate the simulation.