Check for chemgraph redundancy when reading in kinetics_library
When reading in the species.txt file from any kinetics_library folder,
RMG now checks if the species (i.e. chemgraph) was already read-in.  If
so, RMG prints the two species' names (as defined by the user) and
terminates immediately.

Before, RMG would not make a new Species, but would add a new entry
to the localSpecies LinkedHashSet.  The entry would then have the same
species twice, with different names.