Allowing RMG to recognize A.Vandeputte "substitutionS" rxn family template
A.Vandeputte has a "thermo_consistence" (the forward template is equal to the
reverse template) rxn family on his local RMG.  He stated that RMG printed
many warning messages regarding "cannot find rate constant" (see Issue #162).

RMG has hard-coded fixes in the source code to get around this (e.g.
H_Abstraction, intra_H_migration, etc.).  This commit adds the "substitutionS"
rxn family to the list of "thermo_consistence" families that RMG checks
when estimating the kinetics.

This is certainly not the best way to solve the problem, but will allow
A.Vandeputte to use RMG for his purposes in the meantime.  Also, I push it
to the master branch, so that this code is released in future versions of
RMG (and A.Vandeputte will have it automatically).