Added pruning statistics to log.
This prints the number of species pruned, the memory used before and
after pruning, and the memory recovered as a result of pruning. A
warning is printed if the memory use does not decrease as a result of
pruning and more than 100 species are pruned. (Despite invoking garbage
collection before both memory queries, it seems that the memory use
might go up when only a few species are pruned, perhaps due to the
memory used by the local variables of the pruneReactionModel() method.)

These statistics are only printed if one or more species are pruned.