Allow forbidden structures iff they are in a primary thermo library.
Sometimes forbidden structure template will forbid a species that you want,
but you don't want to remove the forbidden structure template in general
because it stops a lot of species that you don't want. With this change,
if a species is in one of your primary thermochemistry libraries, then it
is allowed through, with a warning message.