Updating dictionary and tree for thermo_group/Radical
Three instances in the Radical_Dictionary.txt have a "Cd" functional
group bonded to an Oxygen.  In RMG, "Cd" implies the double bond is to
another Carbon.  This commit changes the "Cd" to "CO".

Since "Cd" and "CO" are mutually exclusive, the updated nodes are no
longer children of their previous parents.  This commit places the
updated nodes one level higher in the tree structure (now a sibling of
their former parent).