Fix problem with forbidden structures messing up thermo estimation.
In 94bf9cb08b82669070d2060be1e7690685719b73 I introduced a feature where if
a forbidden structure is in a primary thermo library then it is allowed.

If a structure was forbidden because it leads to endless stack overflows in
the GATP thermo estimator, then generating the thermo before deciding
whether to forbid the structure is a bad idea.

Now, instead of generating the thermo and seeing if it came from
a primary thermo library, we just search the primary thermo library.