Use isFromPrimaryKineticLibrary() when deciding whether safe to prune reaction Structure.
When preparing a reaction for pruning, we usually clear its Structure.
We don't want to do this if it's in a primary reaction library though,
or we'll mess up the library. This used to be checked for by looking
for "Library" in the kinetics source string. This fails with a null
pointer exception when there is no kinetics (eg PDep Reactions).

Now, we use the isFromPrimaryKineticLibrary() to decide.

For now I have left the old string-checking method in as a sanity check on
the isFromPrimaryKineticLibrary() result. I'm not sure why we needed to check
the reverse direction, but it was there in the code.