Add the ReactionTemplate tracking to PDepReaction objects.
I would prefer it if only TemplateReaction objects had ReactionTemplates,
but if we don't know which ReactionTemplate a reaction came from, we cannot
remove it from the dictionaryByStructure.

This dictionary is now a WeakHashMap, but the reactions in the values, have
reverses, which have structures, which are in other dictionaries. The circular
references across dictionaries keeps it alive. One could break the references
by setting a reaction's reverse to null, but we don't have the reaction object
at the time that we're pruning - we only have a PDepReaction version with
the same structure.

This is the only fix I can think of while maintaining the dictionaries as they
function, and I don't want to risk breaking something I don't understand.