PDepReaction only stores PDepReverse not reverseReaction
The accessor method .getReverseReaction() has been overridden in
PDepReaction class so that it is impossible to access or set the
reverseReaction attribute. Now, it is initialized to null, and should
remain so. If you try to setReverseReaction with a NON PDepReaction
it throws an exception instead of quietly ignoring you.