Pruning of PDep path reactions cope with null reverse
We now get the reverse structure by generating it if there isn't a 
reverse reaction object for us to use. Removing the reverse structure
from the Template dictionary then setting the 'reverse' object to be 
a TemplateReaction, means that when we call reverse.prune() it tries
to remove itself *again* from the dictionary, and fails. This has
been replaced with a warning, but it's still obnoxious.