Made removeFromAllReactionDictionariesByStructure method in ReactionTemplateLibrary
This accepts a reaction Structure, and iterates through all the ReactionTemplate
objects trying to remove the Structure from the dictionaries. This was done for 
PDep Path Reactions in the previous commit, but has now been generalized.

This is called instead of just removing the Structure from the ReactionTemplate
to which its reaction belongs. (in the TemplateReaction.prune() method, which has
now been removed).

Benefit of this approach: when a reaction does not have exactly one Template
(eg. currently PDepReactions have 0 and in the future all Reactions could have
more than 1!) we search all RectionTemplates.

Drawback: we have to search all ReactionTemplates instead of just one per 
TemplateReaction.