Fix problem with Forbidden Structures that are allowed due to Primary Thermo Library.
I'm not sure why this little piece of code is causing me so much trouble,
but this is the second or third bug fix in this vicinity.
Hopefully it's the last. This one was deceptive because some error
handling elsewhere was hiding the real cause by catching all Exceptions
and not printing the stack trace. (will be fixed in separate commit)