Whenever catching a generic Exception e, call Logger.logStackTrace(e).
This should make debugging things a little quicker!
In general it's bad practice to catch generic Exceptions - because
you have no idea what type of exception they are and if you can
really recover from them. At least this way, the stack trace will
be saved in the log file and printed, so that we can quickly find
what actually caused the exception. If it turns out we're logging
lots of exceptions after this commit, then we need to change
the relevant code so that we don't use Exceptions as the norm
(they should be exceptional!)