Remove comments from primary library names
When reading in the name of any primary library, RMG used to take all
characters after the "Name: " field as the name.  However, a user may
include comments in this section which are not intended to be part of
the library name.

RMG still reads in all characters after the "Name: " field, but now
checks whether any comments, i.e. "//", are present.  If so, RMG only
stores the characters before the comment as the library name.

Closes #167