Do not increase redundancy for kinetics supplied via primary kinetic library
This commit addresses Issue #174.

Before: RMG would increase the Arrhenius A factor of user-supplied kinetics
(e.g. from a primary kinetic library) by the structure's redundancy (and
apparently, by other factors as well, since Nick points out his A was
increased by 2^10).  The desired effect is for RMG to use the A factor as
written by the user.

Now: This commit resolves the issue.  When the "addAdditionalKinetics"
function is called, an additional boolean is passed.  This boolean states
whether the addAdditionalKinetics function is being called while reading in
user-specified information or not.

If RMG is reading in user-specified information, we want the function to behave
as before (i.e. if there are multiple kinetics given for a single structure,
include these in the Kinetics vector).  If RMG is not reading in user-specified
information, we only want to proceed with this function if the kinetics are
not user-specified (i.e. primary kinetics library, reaction library, Restart,
etc.).  In this way, the A factor will increase by the redundancy for
RMG-generated kinetics, but not for user-specified kinetics.

Closes #174.