Added new rate to 2+2_cycloaddition_CO rxn family
MRH computed the kinetics for the reaction *CH2-cyclc(CH-CH2-O-O) =>
[ *CH2-O-O-CH=CH2 ] ==> CH2O + CH2CHO.  Visualizing the TS shows the
motion of the C-C bond breaking ... MRH is not sure if the TS also
corresponds to the O-O bond breaking simultaneously ... and has not run
a IRC job to confirm.

However, when MRH tried to get a stable geometry for *CH2-O-O-CH=CH2,
it would always fall apart to CH2O + CH2CHO (at the B3LYP/6-31G(d) level).
The rank I am associating with this rate is 5 (least uncertain).

I added a new node (CH2CHO) to the tree.  RMG used to fall back to
R1-C(=O)-R2 because of the radical on CH2CHO.