Fixing node definition in H_Abstraction rxn family
The node definition for InChI=1/C3H5/c1-3-2/h3H,1-2H2 was causing the
error message: "ERROR: InChI=1/C3H5/c1-3-2/h3H,1-2H2 is not actually a
child of C_rad/H2/Cd".

The *3 label was added to Carbon 1 of the node definition and the 2
hydrogen atoms bonded to Carbon 1 are now explicitly stated (in line with
the definitio of C_rad/H2/Cd).