A simple client/server interface to PopulateReactions
This is a step towards running PopulateReactions as a web service.
I was hoping we could just load all the databases etc. (the slow part) once
and then just send the new reactive species as the request and run them.

It sort of works, but the species and reactions are remembered by the rmg
core-edge reaction model from one run to the next, which messes things up
a little. Perhaps all the appropriate lists/maps/sets can be cleared, or else
the repeating loop (run for each request) will have to be extended to include
all the database loading etc. (this could at least be done in advance of a
request arriving, making it appear to respond more quickly)