PopulateReactions[Server] is now honest about its Ea units (which are kcal/mol)
  The master branch commit 1ab515c6cbd6e8a5538d57346393b21c4815f8a6 made a fix
  to PopulateReactions.java. Unfortunately, PopulateReactionsServer.java is
  95% an exact copy (sorry!) so this has to be changed too. This commit
  makes the changes to both.

Previously it would claim to be in cal/mol, although it was usually in kcal/mol.
The path reactions in the comments above a PDep network would be in the reported
units, because it used the kinetics.toChemkinString() method. The non-dep rates
however were created by a separate function (strangely called updateListeOfReactions)
which ignored the units specification and just printed the E value.
This function has been renamed getFormattedKinetics and has been modified
to use the kinetics.toChemkinString() method. This should respect the
ArrheniusKinetics.setEaUnits() call, which is now "kcal/mol". A side-effect
is that the formatting is now neater as well.