If unnamed species is in a Primary Thermo Library, use the name from that library.
If a species is not named in the input file, but is named in a primary thermo library,
and that name is not an ugly PrIMe ID ("s01234567") then use the name from
the primary thermo library instead of the chemical formula.

This will change the apparent output of some RMG jobs when you change the
primary thermo libraries. The inconsistency is a minor inconvenience, but I
think having descriptive names instead of just the atom counts (molecular formula)
is an improvement...?