More (and different) debugging for issue #185
In Reaction.getReactants(), instead of catching all exceptions, 
we only catch RuntimeExceptions (which includes NullPointerExceptions),
then instead of returning null we throw it again after printing 
the debugging lines.

In the pruning code, if we have a NullPointerException when inspecting 
a path reaction for prunability, then we log more error messages and 
carry on, assuming it's not prunable.