FAME now accepts exponential down parameter as power function.
In recent literature it is common to account for the temperature dependence
of the exponential down parameter with a simple power law of the form

dEdown = alpha * (T / T0)^n

Often the value of n is 0.8-1.0, which implies a pretty significant
temperature effect. This commit enables FAME to read and use the
three-parameter power law expression above.