Use new DeltaEDown class to represent the exponential down parameter.
Note that I have removed averaging the species and bath gas values for
this parameter, as I don't know the best way to do this for a pair of
power law expressions. (The previous approach used an arithmetic mean,
which probably wasn't a great way to do it either. At any rate, the
temperature dependence is more important than the averaging.)