Modifications made to incorporate Sandeep's rates for intra_H migration in HOOQOO species
These are taken from DOI: 10.1021/jp9098792

The A factor in the paper differs from the ones added to the library as the
paper assumes a slightly different Arrhenius form compared to the default RMG
one. A few new groups were added to the dictionary and tree files to account
for the change in the TS ring size or the H being abstracted. The new library
entries have been checked by performing PopulateReactions jobs with the
original molecules in Sandeep's paper and checking if the right entry is
located.