TDE extended to set atom constraints for candidate molecules
The input file of the TDE now has a section were atom constraints,
e.g. max. # of carbons per species etc..., can be set.

new TDE method is based on ChemGraph.readAtomTypes method

When a species does not fulfill the atom constraints, a error will
be thrown and TDE will crash in an unelegant way. This can be fixed
more elegantly.

END keyword has to end the atom constraints section of the input file

example input from /examples/ThermoDataEstimator is updated as well.

Signed-off-by: Nick Vandewiele <nickvandewiele@gmail.com>