PopulateReactionsServer is much more robust. Reports errors rather than crashing.
First problem was that the Logger had not been initialized, so there was no 
log file, so calls to Logger.error() cause it to die.

Second, was that malformed input would kill the server instead of just that 
request. Now the whole servery loop is wrapped in a try/catch block with 
some exception handling.