Logger shouldn't cause crashes when asked to log a warning or error.
It used to crash if the log file had not been initialized. This was the case
in various ancillary classes like PopulateReactions etc. It is still preferable
to call Logger.initialize() at the start of these classes, but if you don't 
at least the logger will no longer cause your code to crash.

It used to cause things like:
Exception in thread "main" java.lang.NullPointerException
    at jing.rxnSys.Logger.flush(Logger.java:120)
    at jing.rxnSys.Logger.log(Logger.java:191)
    at jing.rxnSys.Logger.error(Logger.java:213)