Backported better method of choosing reservoir cutoffs in RS method.
Previously we were choosing the cutoff to be a few dEdown below the lowest
transition state barrier for each unimolecular isomer. In reality, the
energy grains between that cutoff and the transition state barrier aren't
really in pseudo-steady state either, so we might as well place the cutoff
at the barrier height exactly. (We made this change in MEASURE quite a
while ago, so I'm fairly confident that it won't harm anything.)

The result is that the sensitivity of the k(T,P) values to the dEdown is
significantly diminished for the RS method.