Added check that computed k(T,P) values satisfy thermo consistence.
This is a sanity check recently added to RMG-Py to help tease out some
PDep bugs. I feel it is a good idea in RMG-Java as well, although probably
less likely to be triggered due to the network expansion algorithm used
by RMG-Java. This check is necessarily done in Fortran, as FAME no longer
returns both the forward and reverse k(T,P) values.