Multiple distinct ring corrections are taken into account
Before this commit, molecules with multiple rings, would only
get 1 ring correction assigned. For unfused rings, such as 3-cyclo-
pentyl-cyclopentene, two distinct ring corrections, i.e. a cyclopentane
and a cyclopentene RC would be appropriate.

Therefore, ThermoGAGroupLibrary.findRingCorrection() is modified.
This is done by returning a Set of ThermoGAValues, rather than a
single ThermoGAValue.

In terms of the alogorithm, a RC is added to the (returned) Set if and
only if it has the same 'depth' as the deepest Stack that is found so far.
This implies that for now, all RC's should have the same depth (i.e. L2),
otherwise trouble is ahead.

Because a ThermoGAValue is added to the Set of unique elements, an
equals(Object o) method is added to the FunctionalGroup type, comparing
the names of both FunctionalGroups.

Because a Set is returned, a new plus(Set<ThermoGAValue>) mehthod
in ThermoData is created.

Finally, because now multiple distinct RCs are returned I renamed the
methods to getRingCorrections, findCorrections, etc... (instead of
the singular name.

One flaw: a molecule with two identical RCs, e.g. cyclopentyl-cyclopentane,
i.e. two cyclopentane RCs, will currently not be accounted for. This
is because a Set only takes unique RCs.