Multiple identical Ring straings are recognized
In the previous commit, only distinct RSCs could be accounted for,
because of the nature of the Set type, not allowing duplicate keys.

Now, a Map type is used in which the RSC is mapped to the number of
times it has been identified in the Graph.

Signatures of all corresponding methods are modified.

In terms of the algorithm, the following changes are applied:
-instead of iterating over all the Nodes of the Graph, we now iterate
over only the ring nodes. Two separate loops, one for the SSSRings,
and one for the Nodes inside a SSSRing. A List of Sets of RingNodes are
fetched by a new method in the Graph type, called .getCycleNodes().

Furthermore, to anticipate inaccurate predictions of fused ring systems
using single ring strain corrections, fused ring atoms are perceived
by a new method in Graph: .getFusedRingAtoms(). A comment is appended
to the ChemGraph, warning for the presence of fused ring systems.

Fused ring atoms are perceived by first retrieving a List of Sets of Ring
Nodes (cf. supra), and then iteratively checking if Nodes are present
in multiple Sets of RingNodes. These fused ring atoms are returned as a
List.

As a side comment, I believe ring perception algorithms should not
be part of the Graph type, but be externalized, to keep the
core types of RMG (Species, ChemGraph, Graph, etc...) as lightweight as
possible...