WriteRestart added after model enlargement has terminated
Before, the WriteRestart feature would be called in the
model enlargement step, in the loop in ReactionModelGenerator.
modelGeneration():

" while (!allTerminated || !allValid) {"

However, if RMG decides that all fluxes are below the termination
tolerance, the model enlargement is not carried out, and the Restart
folder is not written.

This is annoying, especially when compiling multiple Seed mechanisms,
in which species with multiple nomenclatures, but the same
adjacency list, are used.