Adding R_Addition_CSm family to database.
This commit adds a new reaction family to the RMG database,
which is analogous to the R_Addition_COm family. It is important
to note that the new database essentially is EXACTLY THE SAME
as R_Addition_COm, except that all references to CO are changed
to CS. All rate library data were calculated for carbon monoxide,
and are almost certainly wrong for this case. Thus, they will be
kept as a first estimate, and will be updated in the future.