Chebyshev to PLOG converter (Item 15c from 31Aug2011 RMG meeting)
This function takes a chem.inp file and converts all CHEB type rate coefficients
to a single modified Arrhenius format, fit to the user-specified pressure.

The CHEB information for each reaction is read-in from the chem.inp file.  The
k(T,P) are calculated for 8 temperatures (the number 8 is presently hard-coded)
at the pressure of interest; the temperatures span the TMIN-TMAX range
read-in from the CHEB auxillary information, evenly spaced on an inverse
temperature scale.  These 8 k(T,P) are then fit to a modified Arrhenius format
and these values (and representation) replace the CHEB representation in the
chem.inp file.