Fixed deallocation error when FAME fails due to inconsistent k(T,P) thermo.
When this error occurs, we goto the end of the program, where we clean up
dynamically allocated arrays and then quit. However, if inconsistent thermo
is encountered, the chebyshevCoeffs and pDepArrhenius arrays are never
allocated, which causes an error when we try to deallocate them. Now we
check that they are first allocated before attempting to deallocate them.

Unfortunately this does not fix the inconsistent thermo error that is the
real issue.