Added sanity checking for generated k(E) values to FAME.
These checks ensure that the k(E) values are consistent with the k(T) and
Keq(T) values that were used to compute them. Note that, if they are not
consistent, a warning is printed to FAME.log, but the calculation is not
stopped. If you see this warning at low T but not high T, try increasing
the number of grains or decreasing the grain size. (This makes the case
that the variable number of grains that MEASURE now uses was actually worth
the effort to implement. However, it's complicated enough that I'd rather
not port it back to FAME.)

Unfortunately, this also does not fix the inconsistent k(T,P) thermo bug,
as all of the new warnings disappear when the number of grains is
increased, but the bug remains.