Fixed bug in determination of reservoir cutoffs in RS method in FAME.
For some reason the previous algorithm could lead to cutoffs below the
ground state energy for the case where there is a zero barrier adjacent
to the well. This was apparently causes a singular matrix error in the RS
method (probably fairly frequently). The new algorithm ensures that the
reservoir cutoff is at least equal to the ground-state energy, which seems
to avoid the singular matrix issue.