Backported barrierless reaction k(E) fix to FAME.
As discovered recently in MEASURE, it's possible for barrierless reactions
to produce a kf(E) that is nonzero but a kr(E) that is zero do to numerical
discretization of energies. When this occurs we force both directions to
be zero. As stated in the comments, this is correct to the accuracy of our
energy discretization, and is required to achieve the correct equilibrium
ratios in the generated k(T,P) values.