Don't crash in makeSpeciesIncluded() if there is nothing to do.
When multiple reaction systems are present, multiple species can be added
to the core and/or make included in networks at a time. Depending on the
order, the later actions in this sequence may no longer be valid, e.g. due
to partial network merging. Thus I think it's okay to have
makeSpeciesIncluded() do nothing if it can't find anything to do. A warning
message is printed, just in case.

This is a squashing of issue #210.