In FAME, only allocate Chebyshev or PLOG arrays if used.
Otherwise the Chebyshev number of temperature and pressure polynomial terms
is uninitialized, which can lead to strange behavior (e.g. memory
allocation errors). Since we test for allocated arrays before deallocating
them, this is okay to do. (It should save a bit of memory, too, though
probably not a significant amount compared to other things.)