Compute degeneracy for backward reactions before pdep in PopulateReactions.
This is a copy-paste of similar code that is called just before each
reaction is written to the PopulateReactions output file. That approach
works fine when pdep is off, but we need to do it sooner if pdep is on.
For the moment I've left the original code as well (since PopulateReactions
is not time or memory limited), but it would be better to remove it once we
know that things are working.

(Hopefully) closes #217.