debugged Mike's aromaticity perception algorithm
The addition re-iterates over all nodes in the chemgraph,
with perceived B-bonds, and perceives the atom types again.

If this is not done, the chemgraph will not have the correct atom types
and GAVs thermochemistry will not be correctly perceived.

E.g. aromatic carbons in toluene will still match Cds instead of Cb,etc.

One of the issues why this bug was difficult to track is that the
printed adjacency list only tracks bond information ("B" bond),
and not atom information, in the form of an attribute that could
be as "supplementary information".

Credits of Aromaticity Perception algorithm go to Michael Harper.